@@ -21,26 +21,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.2",
-    "@angular/cdk": "17.1.2",
-    "@angular/common": "^17.1.2",
-    "@angular/compiler": "^17.1.2",
-    "@angular/core": "^17.1.2",
-    "@angular/forms": "^17.1.2",
-    "@angular/localize": "^17.1.2",
-    "@angular/material": "17.1.2",
-    "@angular/material-moment-adapter": "^17.1.2",
-    "@angular/platform-browser": "^17.1.2",
-    "@angular/platform-browser-dynamic": "^17.1.2",
-    "@angular/router": "^17.1.2",
+    "@angular/animations": "^18.2.2",
+    "@angular/cdk": "18.2.2",
+    "@angular/common": "^18.2.2",
+    "@angular/compiler": "^18.2.2",
+    "@angular/core": "^18.2.2",
+    "@angular/forms": "^18.2.2",
+    "@angular/localize": "^18.2.2",
+    "@angular/material": "18.2.2",
+    "@angular/material-moment-adapter": "^18.2.2",
+    "@angular/platform-browser": "^18.2.2",
+    "@angular/platform-browser-dynamic": "^18.2.2",
+    "@angular/router": "^18.2.2",
     "@ng-select/ng-select": "^12.0.7",
-    "@ngrx/component-store": "^17.1.0",
-    "@ngrx/effects": "^17.1.0",
-    "@ngrx/entity": "^17.1.0",
-    "@ngrx/eslint-plugin": "^17.1.0",
-    "@ngrx/router-store": "^17.1.0",
-    "@ngrx/store": "^17.1.0",
-    "@ngrx/store-devtools": "^17.1.0",
+    "@ngrx/component-store": "^18.0.2",
+    "@ngrx/effects": "^18.0.2",
+    "@ngrx/entity": "^18.0.2",
+    "@ngrx/eslint-plugin": "^18.0.2",
+    "@ngrx/router-store": "^18.0.2",
+    "@ngrx/store": "^18.0.2",
+    "@ngrx/store-devtools": "^18.0.2",
     "@progress/kendo-angular-buttons": "16.2.0",
     "@progress/kendo-angular-common": "16.2.0",
     "@progress/kendo-angular-dateinputs": "16.2.0",
@@ -69,36 +69,37 @@
     "imask": "^6.6.0-alpha.0",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
-    "ngrx-store-localstorage": "16.1.0",
+    "ngrx-store-localstorage": "18.0.0",
     "ngx-cookie": "^6.0.1",
     "performant-array-to-tree": "^1.11.0",
     "rxjs": "~7.5.0",
     "tinymce": "^6.8.3",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.3"
+    "zone.js": "~0.14.10",
+    "@ngrx/operators": "^18.0.0"
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^17.0.0",
-    "@angular-devkit/build-angular": "^17.1.2",
-    "@angular-eslint/builder": "17.2.1",
-    "@angular-eslint/eslint-plugin": "17.2.1",
-    "@angular-eslint/eslint-plugin-template": "17.2.1",
-    "@angular-eslint/schematics": "17.2.1",
-    "@angular-eslint/template-parser": "17.2.1",
-    "@angular/cli": "~17.1.2",
-    "@angular/compiler-cli": "^17.1.2",
+    "@angular-devkit/build-angular": "^18.2.2",
+    "@angular-eslint/builder": "18.3.0",
+    "@angular-eslint/eslint-plugin": "18.3.0",
+    "@angular-eslint/eslint-plugin-template": "18.3.0",
+    "@angular-eslint/schematics": "18.3.0",
+    "@angular-eslint/template-parser": "18.3.0",
+    "@angular/cli": "~18.2.2",
+    "@angular/compiler-cli": "^18.2.2",
     "@cypress/code-coverage": "^3.12.20",
     "@cypress/schematic": "2.4.0",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@openapitools/openapi-generator-cli": "^2.5.2",
     "@types/lodash": "^4.14.191",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
-    "@typescript-eslint/parser": "^6.21.0",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "babel-plugin-istanbul": "^6.1.1",
     "cypress": "^13.6.4",
-    "eslint": "^8.56.0",
+    "eslint": "^8.57.0",
     "ng-extract-i18n-merge": "^2.10.0",
     "rimraf": "^3.0.2",
-    "typescript": "~5.3.3"
+    "typescript": "~5.5.4"
   }
 }
