@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
-import { ComponentStore, tapResponse } from '@ngrx/component-store';
-import { concatLatestFrom } from '@ngrx/effects';
+import { ComponentStore } from '@ngrx/component-store';import { tapResponse, concatLatestFrom } from '@ngrx/operators';
+
+
 import { Store } from '@ngrx/store';
 import { Observable, mergeMap } from 'rxjs';
 import {
