@@ -1,5 +1,6 @@
 import { Injectable, OnDestroy } from '@angular/core';
-import { ComponentStore, tapResponse } from '@ngrx/component-store';
+import { ComponentStore } from '@ngrx/component-store';import { tapResponse } from '@ngrx/operators';
+
 import { Observable, map, mergeMap } from 'rxjs';
 import { APIItContractResponseDTO, APIV2ItContractService } from 'src/app/api/v2';
 import { filterNullish } from 'src/app/shared/pipes/filter-nullish';
