@@ -1,8 +1,10 @@
-import { Injectable } from '@angular/core';
-import { Actions, concatLatestFrom, createEffect, ofType } from '@ngrx/effects';
+import { Inject, Injectable } from '@angular/core';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
+import { concatLatestFrom } from '@ngrx/operators';
+
 import { Store } from '@ngrx/store';
-import { catchError, filter, map, of, switchMap } from 'rxjs';
-import { APIV2OrganizationService } from 'src/app/api/v2';
+import { catchError, combineLatestWith, filter, map, of, switchMap } from 'rxjs';
+import { APIV2OrganizationService, APIV2OrganizationUnitsInternalINTERNALService } from 'src/app/api/v2';
 import { BOUNDED_PAGINATION_QUERY_MAX_SIZE } from 'src/app/shared/constants';
 import { filterNullish } from 'src/app/shared/pipes/filter-nullish';
 import { selectOrganizationUuid } from '../user-store/selectors';
@@ -14,7 +16,9 @@ export class OrganizationUnitEffects {
   constructor(
     private actions$: Actions,
     private store: Store,
-    private apiOrganizationService: APIV2OrganizationService
+    @Inject(APIV2OrganizationService) private apiOrganizationService: APIV2OrganizationService,
+    @Inject(APIV2OrganizationUnitsInternalINTERNALService)
+    private apiUnitService: APIV2OrganizationUnitsInternalINTERNALService
   ) {}
 
   getOrganizationUnits$ = createEffect(() => {
@@ -46,6 +50,25 @@ export class OrganizationUnitEffects {
               return OrganizationUnitActions.getOrganizationUnits(maxPageSize, nextPage, allUnits);
             }),
             catchError(() => of(OrganizationUnitActions.getOrganizationUnitsError()))
+          )
+      )
+    );
+  });
+
+  patchOrganizationUnit$ = createEffect(() => {
+    return this.actions$.pipe(
+      ofType(OrganizationUnitActions.patchOrganizationUnit),
+      combineLatestWith(this.store.select(selectOrganizationUuid).pipe(filterNullish())),
+      switchMap(([{ unitUuid, request }, organizationUuid]) =>
+        this.apiUnitService
+          .patchSingleOrganizationUnitsInternalV2PatchUnit({
+            organizationUuid,
+            organizationUnitUuid: unitUuid,
+            parameters: request,
+          })
+          .pipe(
+            map((unit: any) => OrganizationUnitActions.patchOrganizationUnitSuccess(unit)),
+            catchError(() => of(OrganizationUnitActions.patchOrganizationUnitError()))
           )
       )
     );
